PySpice_studio/requirements.txt
schemdraw
matplotlib
Pillow
numpy